requirements.txt
black==24.10.0
click==8.1.7
flake8==7.1.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pycodestyle==2.12.1
PyEventEmitter==1.0.5
pyflakes==3.2.0
PySide6==6.8.0.2
PySide6_Addons==6.8.0.2
PySide6_Essentials==6.8.0.2
shiboken6==6.8.0.2
tomli==2.0.2
typing_extensions==4.12.2
